Successfully reading Accel and Gyro @ 100 HZ. If sped up there are issues keeping everything synchronous.
Had 1 bug where the Accel was not writing the address it wanted to read from which would cause the Accel data transfer to hang. Ended up having to move I2C__MultReadStart() to a timer. I believe that there was a race condition between 2 interrupts. I2C interrupt and gpioD since both were calling 'startmultread'

Still need to speed up the comm with both ac & gy so that we can filter properly.
Also will need to synchronize the accell and gyro samples. they are currently asychronous which will make the filter behavior unpredictable